Fix an oversight in set iterator constructor
The bit twiddling operation would clear all bits _below_ the slot index given - if given slot_index 1 would clear bit 0 but not bit 1, which is not what is needed. We need to mask those out but also clear the bit for slot_index. Doing it as shown here prevents UB if, say, mask_type is uint32_t and slot_index is 31.